--- feature/src/main/java/com/zcard/feature/cardeditor/model/BaseTabItem.kt
@@ -1,6 +1,0 @@
-package com.zcard.feature.cardeditor.model
-
-data class BaseTabItem(
-    val id: String,
-    val textRes: Int
-)
